update float to double in lattice.cpp

+++ src/Lattice.cpp
@@ -573,7 +573,7 @@ void Lattice::generateCSGLists(std::vector<int>* surf_flags, std::vector<double>
 	/* Loop over all lattice cells from left to right, bottom to top */
 	for (int i=0; i < _num_y; i++) {
 
-		current_origin->setCoords(current_origin->getX(), global_y + (i * _width_y) - ((float(_num_y)/2.0 - .5) * _width_y));
+		current_origin->setCoords(current_origin->getX(), global_y + (i * _width_y) - ((double(_num_y)/2.0 - .5) * _width_y));
 
 
 		for (int j=0; j < _num_x; j++) {
@@ -592,7 +592,7 @@ void Lattice::generateCSGLists(std::vector<int>* surf_flags, std::vector<double>
 			Cell* cell = cells.begin()->second;
 
 			/* update current_origin */
-			current_origin->setCoords(global_x + (j * _width_x) - ((float(_num_x)/2.0 - .5) * _width_x),
+			current_origin->setCoords(global_x + (j * _width_x) - ((double(_num_x)/2.0 - .5) * _width_x),
 					current_origin->getY());
 
 			log_printf(DEBUG, "Lattice: set current_origin: (%f,%f)", current_origin->getX(), current_origin->getY());
